+++ modules/canbus/vehicle/transit/protocol/adc_motioncontrollimits1_12.cc
@@ -63,10 +63,10 @@ void Adcmotioncontrollimits112::set_p_adc_cmd_throttlecommandlimit(
     uint8_t* data, double adc_cmd_throttlecommandlimit) {
   adc_cmd_throttlecommandlimit =
       ProtocolData::BoundedValue(0.0, 100.0, adc_cmd_throttlecommandlimit);
-  int x = adc_cmd_throttlecommandlimit / 0.500000;
+  int x = static_cast<int>(adc_cmd_throttlecommandlimit / 0.500000);
 
   Byte to_set(data + 3);
-  to_set.set_value(x, 0, 8);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 8);
 }
 
 Adcmotioncontrollimits112* Adcmotioncontrollimits112::set_adc_cmd_steeringrate(
@@ -83,15 +83,15 @@ void Adcmotioncontrollimits112::set_p_adc_cmd_steeringrate(
     uint8_t* data, double adc_cmd_steeringrate) {
   adc_cmd_steeringrate =
       ProtocolData::BoundedValue(0.0, 3276.75, adc_cmd_steeringrate);
-  int x = adc_cmd_steeringrate / 0.050000;
+  int x = static_cast<int>(adc_cmd_steeringrate / 0.050000);
   uint8_t t = 0;
 
-  t = x & 0xFF;
+  t = static_cast<uint8_t>(x & 0xFF);
   Byte to_set0(data + 0);
   to_set0.set_value(t, 0, 8);
   x >>= 8;
 
-  t = x & 0xFF;
+  t = static_cast<uint8_t>(x & 0xFF);
   Byte to_set1(data + 1);
   to_set1.set_value(t, 0, 8);
 }
@@ -112,10 +112,10 @@ void Adcmotioncontrollimits112::set_p_adc_cmd_steerwheelanglelimit(
     uint8_t* data, double adc_cmd_steerwheelanglelimit) {
   adc_cmd_steerwheelanglelimit =
       ProtocolData::BoundedValue(0.0, 1275.0, adc_cmd_steerwheelanglelimit);
-  int x = adc_cmd_steerwheelanglelimit / 5.000000;
+  int x = static_cast<int>(adc_cmd_steerwheelanglelimit / 5.000000);
 
   Byte to_set(data + 2);
-  to_set.set_value(x, 0, 8);
+  to_set.set_value(static_cast<uint8_t>(x), 0, 8);
 }
 
 }  // namespace transit
